LV: Remove circular dependency
